+++ R/model_default.R
@@ -1,3 +1,84 @@
+#' Ordinary Differential Equations for the Default Model
+#'
+#' @description Provides the ODEs for the default SEIR-V model in a format that
+#' is suitable for passing to [deSolve::lsoda()].
+#' See [model_default()] for a list of required parameters.
+#'
+#' @param t A single number of the timestep at which to integrate.
+#' @param y The conditions of the epidemiological compartments.
+#' @param params The parameters, passed as a named list.
+#'
+#' @return A list with a vector with as many elements as the number of
+#' demographic groups times the number of epidemiological compartments. Each
+#' value gives the change in the number of individuals in that compartment.
+#' @keywords internal
+.ode_model_default <- function(t, y, params) {
+  # no input checking, fn is unsafe and not expected to be used
+  n_age <- nrow(params[["contact_matrix"]])
+
+  # create a matrix
+  y <- matrix(y, nrow = n_age, ncol = 5L, byrow = FALSE)
+
+  # scale the contact matrix if within the intervention period
+  contact_matrix_ <- intervention_on_cm(
+    t = t,
+    cm = params[["contact_matrix"]],
+    time_begin = params[["npi_time_begin"]],
+    time_end = params[["npi_time_end"]],
+    cr = params[["npi_cr"]]
+  )
+
+  # get paramters to modify them
+  # NOTE: `model_params` refers to epdemiological parameters, while
+  # `params` refers to the list passed as a function argument.
+  # e.g. contact matrix, or interventions, are not included in `model_params`
+  model_params <- params[c(
+    "transmission_rate", "infectiousness_rate", "recovery_rate"
+  )]
+
+  # apply time dependence before interventions
+  time_dependent_params <- Map(
+    model_params[names(params$time_dependence)],
+    params$time_dependence,
+    f = function(x, func) {
+      func(time = t, x = x)
+    }
+  )
+
+  # assign time-modified param values
+  model_params[names(time_dependent_params)] <- time_dependent_params
+
+  model_params <- .intervention_on_rates(
+    t = t,
+    interventions = params[["rate_interventions"]],
+    parameters = model_params
+  )
+
+  # modify the vaccination rate depending on the regime
+  # the number of doses is already checked before passing
+  current_nu <- params[["vax_nu"]] *
+    ((params[["vax_time_begin"]] < t) &
+      (params[["vax_time_end"]] > t))
+
+  # calculate transitions
+  sToE <- (model_params[["transmission_rate"]] * y[, 1] *
+    contact_matrix_ %*% y[, 3])
+  eToI <- model_params[["infectiousness_rate"]] * y[, 2]
+  iToR <- model_params[["recovery_rate"]] * y[, 3]
+  sToV <- current_nu * y[, 1]
+
+  # define compartmental changes
+  dS <- -sToE - sToV
+  dE <- sToE - eToI
+  dI <- eToI - iToR
+  dR <- iToR
+  dV <- sToV
+
+  # return a list
+  list(c(dS, dE, dI, dR, dV))
+}
+
+
 #' @title Model an SEIR-V epidemic with interventions
 #'
 #' @name model_default
@@ -119,6 +200,7 @@
 #' )
 #'
 #' data
+#' @importFrom odin odin
 #' @export
 model_default <- function(population,
                           transmission_rate = 1.3 / 7.0,
@@ -139,7 +221,6 @@ model_default <- function(population,
   checkmate::assert_numeric(transmission_rate, lower = 0, finite = TRUE)
   checkmate::assert_numeric(infectiousness_rate, lower = 0, finite = TRUE)
   checkmate::assert_numeric(recovery_rate, lower = 0, finite = TRUE)
-  checkmate::assert_integerish(time_end, lower = 0)
 
   # check the time end and increment
   # restrict increment to lower limit of 1e-6
@@ -252,12 +333,140 @@ model_default <- function(population,
   model_output[, args := apply(model_output, 1, function(x) {
     c(x[["args"]], x[param_names]) # avoid including col "param_set"
   })]
-  model_output[, "data" := Map(population, args, f = function(p, l) {
-    .output_to_df(
-      do.call(.model_default_cpp, l),
-      population = p, # taken from local scope/env
-      compartments = compartments
+
+  model_output[, "data" := lapply(args, function(args) {
+    time_points <- seq(0, args$time_end, by = args$increment)
+    n_time <- length(time_points)
+
+    C <- args$contact_matrix
+    n_age <- nrow(C)
+
+    contact_intervention_start <- as.numeric(args$npi_time_begin)
+    contact_intervention_end <- as.numeric(args$npi_time_end)
+    contact_intervention_effect <- t(args$npi_cr)
+
+    rate_intervention_start <-
+      as.numeric(args$rate_interventions[[1]]$time_begin)
+    rate_intervention_end <- as.numeric(args$rate_interventions[[1]]$time_end)
+    rate_intervention_effect <-
+      matrix(rep(args$rate_interventions[[1]]$reduction, n_age), ncol = n_age)
+
+    n_contact_intervention <- length(contact_intervention_start)
+    n_rate_intervention <- length(rate_intervention_start)
+
+    time_dependent_params <- Map(
+      args[names(args$time_dependence)],
+      args$time_dependence,
+      f = function(x, func) {
+        func(time = time_points, x = x)
+      }
     )
+    # assign time-modified param values
+    args[names(time_dependent_params)] <- time_dependent_params
+    beta <- args$transmission_rate
+    sigma <- args$infectiousness_rate
+    gamma <- args$recovery_rate
+
+    if (length(beta) == 1) beta <- rep(beta, n_time)
+    if (length(sigma) == 1) sigma <- rep(sigma, n_time)
+    if (length(gamma) == 1) gamma <- rep(gamma, n_time)
+
+    vax_start <- as.numeric(args$vax_time_begin)
+    vax_end <- as.numeric(args$vax_time_end)
+    vax_nu <- as.numeric(args$vax_nu)
+
+    initial_conditions <- args$initial_state
+    init_S <- initial_conditions[, 1]
+    init_E <- initial_conditions[, 2]
+    init_I <- initial_conditions[, 3]
+    init_R <- initial_conditions[, 4]
+    init_V <- initial_conditions[, 5]
+
+    # Initialize and run the model
+    model <- seirv_model$new(
+      time = time_points,
+      n_time = n_time,
+      C = C,
+      n_age = n_age,
+      n_contact_intervention = n_contact_intervention,
+      n_rate_intervention = n_rate_intervention,
+      beta = beta,
+      sigma = sigma,
+      gamma = gamma,
+      rate_intervention_start = rate_intervention_start,
+      rate_intervention_end = rate_intervention_end,
+      rate_intervention_effect = rate_intervention_effect,
+      contact_intervention_start = contact_intervention_start,
+      contact_intervention_end = contact_intervention_end,
+      contact_intervention_effect = contact_intervention_effect,
+      vax_start = vax_start,
+      vax_end = vax_end,
+      vax_nu = vax_nu,
+      init_S = init_S,
+      init_E = init_E,
+      init_I = init_I,
+      init_R = init_R,
+      init_V = init_V
+    )
+
+    result <- model$run(time_points)
+
+    # Add scenario information
+    dt <- data.table::as.data.table(result)
+    # declaring variables below to avoid data.table related lintr messages
+    temp <- value <- temp_compartment <- temp_demography <-
+      compartment <- demography_group <- `:=` <- time <- NULL
+
+    age_group_mappings <- paste0( # properly label demography groups
+      seq_len(n_age),
+      c(
+        rownames(C),
+        names(population[[1]]$demography_vector),
+        sprintf(
+          "demo_group_%i",
+          seq_len(nrow(population[[1]]$contact_matrix))
+        )
+      )[seq_len(nrow(population[[1]]$contact_matrix))]
+    )
+    names(age_group_mappings) <- seq_len(nrow(population[[1]]$contact_matrix))
+
+    mapping <- c( # prepend numbers to help during sorting. Will remove later
+      S = "1susceptible", E = "2exposed", I = "3infectious",
+      R = "4recovered", V = "5vaccinated", age_group_mappings
+    )
+
+    # Melt the data table to long format
+    data.table::melt(dt,
+      id.vars = "t",
+      variable.name = "temp", # e.g. S[1], ..., V[3]
+      value.name = "value"
+    )[ # piping the data.table way. Possible because melt outputs a data.table
+      , list(
+        time = t, # alternative to using data.table::setnames(dt, "t", "time")
+        temp_compartment = substring(temp, 1L, 1L), # e.g. S[1] -> S
+        temp_demography = substring(temp, 3L, 3L), # e.g. S[1] -> 1
+        value
+      )
+    ][ # |> the DT way (piping the data.table way)
+      ,
+      list(
+        time,
+        demography_group = mapping[temp_demography], # e.g. 1[0,20), 2[20,65),
+        compartment = mapping[temp_compartment], # e.g. 1susceptible, 2exposed
+        value
+      )
+    ][ # |> the DT way
+      order(time, compartment, demography_group) # prepending numbers helps here
+    ][ # |> the DT way
+      ,
+      `:=`( # used as the prefix form to update multiple columns
+        # remove prepended numbers from `mapping`
+        demography_group = substring(demography_group, 2L), # e.g. [0,20), ...
+        compartment = substring(compartment, 2L) # e.g. susceptible, exposed,
+      )
+    ][ # |> the DT way
+      # added because the previous operation used `:=` which doesn't output
+    ]
   })]
 
   # remove temporary arguments
@@ -271,84 +480,4 @@ model_default <- function(population,
 
   # return data.table
   model_output[]
-}
-
-#' Ordinary Differential Equations for the Default Model
-#'
-#' @description Provides the ODEs for the default SEIR-V model in a format that
-#' is suitable for passing to [deSolve::lsoda()].
-#' See [model_default()] for a list of required parameters.
-#'
-#' @param t A single number of the timestep at which to integrate.
-#' @param y The conditions of the epidemiological compartments.
-#' @param params The parameters, passed as a named list.
-#'
-#' @return A list with a vector with as many elements as the number of
-#' demographic groups times the number of epidemiological compartments. Each
-#' value gives the change in the number of individuals in that compartment.
-#' @keywords internal
-.ode_model_default <- function(t, y, params) {
-  # no input checking, fn is unsafe and not expected to be used
-  n_age <- nrow(params[["contact_matrix"]])
-
-  # create a matrix
-  y <- matrix(y, nrow = n_age, ncol = 5L, byrow = FALSE)
-
-  # scale the contact matrix if within the intervention period
-  contact_matrix_ <- intervention_on_cm(
-    t = t,
-    cm = params[["contact_matrix"]],
-    time_begin = params[["npi_time_begin"]],
-    time_end = params[["npi_time_end"]],
-    cr = params[["npi_cr"]]
-  )
-
-  # get paramters to modify them
-  # NOTE: `model_params` refers to epdemiological parameters, while
-  # `params` refers to the list passed as a function argument.
-  # e.g. contact matrix, or interventions, are not included in `model_params`
-  model_params <- params[c(
-    "transmission_rate", "infectiousness_rate", "recovery_rate"
-  )]
-
-  # apply time dependence before interventions
-  time_dependent_params <- Map(
-    model_params[names(params$time_dependence)],
-    params$time_dependence,
-    f = function(x, func) {
-      func(time = t, x = x)
-    }
-  )
-
-  # assign time-modified param values
-  model_params[names(time_dependent_params)] <- time_dependent_params
-
-  model_params <- .intervention_on_rates(
-    t = t,
-    interventions = params[["rate_interventions"]],
-    parameters = model_params
-  )
-
-  # modify the vaccination rate depending on the regime
-  # the number of doses is already checked before passing
-  current_nu <- params[["vax_nu"]] *
-    ((params[["vax_time_begin"]] < t) &
-      (params[["vax_time_end"]] > t))
-
-  # calculate transitions
-  sToE <- (model_params[["transmission_rate"]] * y[, 1] *
-    contact_matrix_ %*% y[, 3])
-  eToI <- model_params[["infectiousness_rate"]] * y[, 2]
-  iToR <- model_params[["recovery_rate"]] * y[, 3]
-  sToV <- current_nu * y[, 1]
-
-  # define compartmental changes
-  dS <- -sToE - sToV
-  dE <- sToE - eToI
-  dI <- eToI - iToR
-  dR <- iToR
-  dV <- sToV
-
-  # return a list
-  list(c(dS, dE, dI, dR, dV))
 }